Change git prompt colors to reflect LoC delta
I'm about to go through a series of refactors, and want a quick visual
indication of whether I'm adding or removing lines of code overall.
This change shows whether I'm red or green from the upstream branch
and from `HEAD`.

![Annotated screenshot of what the new prompt colors represent](http://i.imgur.com/XfNvTuO.png)

It's a bit slow. I'd like to speed it up if this works out.